+++ pdfforms/__init__.py
@@ -1,3 +1,4 @@
 "inspect and fill pdf fillable forms"
-from .pdfforms import comma_format, fill, inspect, round_float
+from .pdfforms import fill_pdfs, inspect_pdfs
+from .transforms import *
 from .version import __version__