+++ CHANGELOG.rst
@@ -1,8 +1,9 @@
 Changelog
------------
+==========
+
 2.0.0
 """""
-:date: 23 June, 2021
+:date: 24 June, 2021
 
 * Use pyexcel to load spreadsheet data, supports xlsx, ods, csv, and more
 * Add options to round values, add thousands separators
@@ -16,7 +17,7 @@ Changelog
 """""
 :date: 3 July, 2020
 
-* Don’t crash when subcommand not supplied. (Thanks @PiDelport for the PR)
+* Don’t crash when subcommand not supplied (Thanks @PiDelport for the PR)
 
 1.2.0
 """""